requirements.txt
numpy
torch
torchvision
matplotlib
git+https://github.com/openai/CLIP.git
Pillow
